requirements.txt
azure-identity==1.19.0
azure-storage-blob==12.24.0
mlflow==2.19.0
notebook==7.3.2
pandas==2.2.3
psycopg2==2.9.10
python-dotenv==1.0.1
tensorflow==2.18.0